% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_bh.R
\name{agg_bh}
\alias{agg_bh}
\title{Aggregate booking horizon}
\usage{
agg_bh(data, bh_len = 30)
}
\arguments{
\item{data}{Numeric vector of demand}

\item{bh_len}{Length of booking horizon. Default 30.}
}
\value{
A numeric vector
}
\description{
This function aggregates multiple time points into a booking horizon
}
