% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emsr_limits.R
\name{emsr_limits}
\alias{emsr_limits}
\title{EMSR Limits}
\usage{
emsr_limits(fares, means, vars, capacity)
}
\arguments{
\item{fares}{Vector of fares.}

\item{means}{Vector of demand mean forecast for each fare class.}

\item{vars}{Vector of demand variance forecast for each fare class.}

\item{capacity}{nteger for total capacity.}
}
\value{
A numeric vector
}
\description{
This function calculates the EMSR booking limits
}
