% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_demand.R
\name{forecast_demand}
\alias{forecast_demand}
\title{Forecast demand}
\usage{
forecast_demand(cust, probs)
}
\arguments{
\item{cust}{Numeric vector of customer arrivals per time period}

\item{probs}{Numeric vector of probability of purchase in each fare class}
}
\value{
A numeric value.
}
\description{
This function forecasts demand per fare class
}
