% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cumulative_counts.R
\name{get_cumulative_counts}
\alias{get_cumulative_counts}
\title{Get cumulative counts}
\usage{
get_cumulative_counts(data)
}
\arguments{
\item{data}{A numeric matrix or dataframe}
}
\value{
A matrix or dataframe
}
\description{
This function applies cumsum to each row of a dataframe or matrix
}
