% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{nhpp}
\alias{nhpp}
\title{nhpp}
\usage{
nhpp(
  alpha = 200,
  beta = 1,
  phi1 = 0.5,
  a1 = 1,
  b1 = 1,
  a2 = 1,
  b2 = 1,
  n = 3000,
  hor_length = 30
)
}
\arguments{
\item{alpha}{Shape parameter of gamma distribution for total demand. Default 200.}

\item{beta}{Rate parameter of gamma distribution for total demand. Default 1.}

\item{phi1}{Percentage of total demand for customer type 1. Default 0.5.}

\item{a1}{Beta distribution parameter 1 for customer type 1. Default 1.}

\item{b1}{Beta distribution parameter 2 for customer type 1. Default 1.}

\item{a2}{Beta distribution parameter 1 for customer type 2. Default 1.}

\item{b2}{Beta distribution parameter 2 for customer type 2. Default 1.}

\item{n}{Total number of time points that will be generated. Default 3000.}

\item{hor_length}{Number of points that make up the booking horizon, Default 30.}
}
\value{
A numeric value.
}
\description{
This function simulates a non-homogeneous Poisson process.
}
