% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpprate.R
\name{nhpprate}
\alias{nhpprate}
\title{nhpprate}
\usage{
nhpprate(A = 240, a = 1, b = 1, phi = 1, t = 1, n = 3000)
}
\arguments{
\item{A}{Mean total demand across all customer types. Default 200.}

\item{a}{Beta distribution parameter 1. Default 1.}

\item{b}{Beta distribution parameter 2. Default 1.}

\item{phi}{Percentage of total demand for this customer type. Default 1.}

\item{t}{Current time point. Default 1.}

\item{n}{Total number of time points that will be generated. Default 3000.}
}
\value{
A numeric value.
}
\description{
This function determines the rate of a non-homogeneous Poisson process.
}
